finish ETRM morris
